+++ processor/src/test/fixtures/bad_input/com/example/fragment/includes/SelfIncludesModel.java
@@ -1,0 +1,17 @@
+package com.example.fragment.includes;
+
+import sting.Fragment;
+import sting.Injectable;
+
+public interface SelfIncludesModel
+{
+  @Fragment( includes = { MyComponent.class, MyInjector.class } )
+  interface MyInjector
+  {
+  }
+
+  @Injectable
+  class MyComponent
+  {
+  }
+}