+++ processor/src/test/fixtures/bad_input/com/example/injector/includes/SelfIncludesModel.java
@@ -1,0 +1,18 @@
+package com.example.injector.includes;
+
+import sting.Injectable;
+import sting.Injector;
+
+public interface SelfIncludesModel
+{
+  @Injector( includes = { MyComponent.class, MyInjector.class } )
+  interface MyInjector
+  {
+    MyComponent getMyComponent();
+  }
+
+  @Injectable
+  class MyComponent
+  {
+  }
+}